+++ core/frontend/helpers/link.js
@@ -1,7 +1,7 @@
 // # link helper
 const _ = require('lodash');
-const {config, SafeString, errors, i18n} = require('../services/proxy');
-const {buildLinkClasses} = require('./utils');
+const {config, SafeString, errors, i18n, localUtils} = require('../services/proxy');
+const {buildLinkClasses} = localUtils;
 
 const managedAttributes = ['href', 'class', 'activeClass', 'parentActiveClass'];
