+++ core/frontend/services/themes/handlebars/template.js
@@ -2,8 +2,8 @@
 const templates = {};
 const _ = require('lodash');
 const errors = require('@tryghost/errors');
-const hbs = require('../services/themes/engine');
-const i18n = require('../../server/lib/common/i18n');
+const hbs = require('../engine');
+const {i18n} = require('../../../../server/lib/common');
 
 // Execute a template helper
 // All template helpers are register as partial view.
